Revert "feat: Migrate to standardized build configs (#1965)" (#1976)
This reverts commit 4219e37dbab7402f3d20133a73f045e228e9f25d.
