--- .changeset/sharp-falcons-glow.md
@@ -1,17 +1,0 @@
----
-"@scalar/use-codemirror": patch
-"@scalar/api-reference": patch
-"@scalar/build-tooling": patch
-"@scalar/object-utils": patch
-"@scalar/mock-server": patch
-"@scalar/use-tooltip": patch
-"@scalar/api-client": patch
-"@scalar/client-app": patch
-"@scalar/components": patch
-"@scalar/use-toasts": patch
-"@scalar/use-modal": patch
-"@scalar/galaxy": patch
-"@scalar/themes": patch
----
-
-Update build configs to a standardized format
