+++ package.json
@@ -1,9 +1,6 @@
 {
   "name": "root",
   "type": "module",
-  "engines": {
-    "pnpm": "^9.1.0"
-  },
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
     "@playwright/test": "^1.42.0",
