+++ packages/api-reference/tsconfig.json
@@ -1,5 +1,11 @@
 {
   "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "paths": {
+      "@/*": ["./src/*"],
+      "@test/*": ["./test/*"]
+    }
+  },
   "include": ["src", "src/declarations", "global.d.ts"],
   "exclude": ["dist", "vite.config.ts", "node_modules"],
   "ts-node": {
