+++ packages/build-tooling/rollup.config.js
@@ -1,0 +1,13 @@
+import typescript from '@rollup/plugin-typescript'
+
+export default [
+  {
+    input: './src/index.ts',
+    output: {
+      file: 'dist/index.js',
+      format: 'esm',
+      sourcemap: false,
+    },
+    plugins: [typescript()],
+  },
+]
