+++ packages/build-tooling/tsconfig.build.json
@@ -14,7 +14,6 @@
     "declaration": true,
     "declarationMap": true,
     "emitDeclarationOnly": true,
-    "outDir": "dist/",
-    "target": "ESNext"
+    "outDir": "dist/"
   }
 }
