+++ packages/components/vite.config.ts
@@ -1,11 +1,35 @@
-import { createViteBuildOptions } from '@scalar/build-tooling'
 import vue from '@vitejs/plugin-vue'
+import * as path from 'path'
 import { defineConfig } from 'vite'
 import dts from 'vite-plugin-dts'
 import svgLoader from 'vite-svg-loader'
 
+import pkg from './package.json'
+
 // https://vitejs.dev/config/
 export default defineConfig({
+  build: {
+    lib: {
+      // Could also be a dictionary or array of multiple entry points
+      entry: './src/index.ts',
+      name: '@scalar/components',
+      formats: ['es', 'cjs'],
+      fileName: 'index',
+      // the proper extensions will be added
+      // fileName: 'my-lib',
+    },
+    rollupOptions: {
+      input: {
+        main: path.resolve(__dirname, 'src/index.ts'),
+      },
+      // make sure to externalize deps that shouldn't be bundled
+      // into your library
+      external: [
+        ...Object.keys(pkg.dependencies),
+        ...Object.keys(pkg.peerDependencies),
+      ],
+    },
+  },
   plugins: [
     vue(),
     dts({ insertTypesEntry: true, rollupTypes: true }),
@@ -27,7 +51,4 @@ export default defineConfig({
       },
     }),
   ],
-  build: createViteBuildOptions({
-    entry: ['src/index.ts'],
-  }),
 })
