+++ packages/mock-server/package.json
@@ -43,7 +43,6 @@
     "hono": "^4.2.7"
   },
   "devDependencies": {
-    "@scalar/build-tooling": "workspace:*",
     "@scalar/galaxy": "workspace:*",
     "@types/node": "^20.8.4",
     "esbuild": "^0.21.1",
