+++ packages/themes/package.json
@@ -48,7 +48,6 @@
   "module": "./dist/index.js",
   "devDependencies": {
     "@rise8/tailwind-pixel-perfect-preset": "^1.0.1",
-    "@scalar/build-tooling": "workspace:*",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.5.0",
     "tailwindcss": "^3.4.3",
