+++ packages/use-codemirror/package.json
@@ -61,7 +61,6 @@
     "codemirror": "^6.0.0"
   },
   "devDependencies": {
-    "@scalar/build-tooling": "workspace:*",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.5.0",
     "tsc-alias": "^1.8.8",
