+++ packages/use-modal/package.json
@@ -42,7 +42,6 @@
   "module": "dist/index.js",
   "devDependencies": {
     "@headlessui/vue": "^1.7.20",
-    "@scalar/build-tooling": "workspace:*",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.5.0",
     "tsc-alias": "^1.8.8",
