+++ packages/use-toasts/package.json
@@ -38,7 +38,6 @@
     "CHANGELOG.md"
   ],
   "devDependencies": {
-    "@scalar/build-tooling": "workspace:*",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.5.0",
     "nanoid": "^5.0.7",
