+++ packages/use-toasts/vite.config.ts
@@ -1,13 +1,24 @@
-import { createViteBuildOptions } from '@scalar/build-tooling'
 import vue from '@vitejs/plugin-vue'
 import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 import { defineConfig } from 'vitest/config'
 
+import pkg from './package.json'
+
 export default defineConfig({
   plugins: [vue(), cssInjectedByJsPlugin()],
-  build: createViteBuildOptions({
-    entry: ['src/index.ts'],
-  }),
+  build: {
+    cssCodeSplit: false,
+    minify: false,
+    lib: {
+      entry: ['src/index.ts'],
+      name: pkg.name,
+      fileName: 'index',
+      formats: ['es'],
+    },
+    rollupOptions: {
+      external: [...Object.keys(pkg.peerDependencies)],
+    },
+  },
   test: {
     coverage: {
       enabled: true,
