+++ packages/use-tooltip/package.json
@@ -41,7 +41,6 @@
   ],
   "module": "dist/index.js",
   "devDependencies": {
-    "@scalar/build-tooling": "workspace:*",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.5.0",
     "tippy.js": "^6.3.7",
