+++ packages/use-tooltip/vite.config.ts
@@ -1,12 +1,23 @@
-import { createViteBuildOptions } from '@scalar/build-tooling'
 import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vitest/config'
 
+import pkg from './package.json'
+
 export default defineConfig({
   plugins: [vue()],
-  build: createViteBuildOptions({
-    entry: ['src/index.ts'],
-  }),
+  build: {
+    cssCodeSplit: false,
+    minify: false,
+    lib: {
+      entry: ['src/index.ts'],
+      name: pkg.name,
+      fileName: 'index',
+      formats: ['es'],
+    },
+    rollupOptions: {
+      external: [...Object.keys(pkg.peerDependencies)],
+    },
+  },
   test: {
     coverage: {
       enabled: true,
